fix: Change approach to removing image tag
The `ParseImage` function processes a docker image name to return
a `ClusterImage` struct instance from it. In the process, the previous
approach to remove the tag is to split the image name by ":"  and to
get the first part. This approach fails when there's more than one ":"
in the image, such as when the registry URL has an explicit port. Because
of this, the function would panic with the following error:

```
index out of range [1] with length 1
```

This commit performs the tag separation **after** the image URL
has been split by "/", only to the last part of the resulting
slice. This way the presence of a port won't affect the function.